Setup a route for handling requests
Setup a route for handling requests which additionally added a logger
when the requests exits.

All the checks will be done in the handle function of the main file
since chaining methods may not be an easy affair.

The next attempt would be to implement something like this
```
handle(...) -> Result<impl Reply, Rejection> {
        pre_1(...);
        pre_2(...);
        ..
        client.handle(...).await;
        post_1(...);
        post_2(...);

        Ok(warp::reply::with_status(reply, http::OK)
}
```